Adds support for a malpedia.json config
Adds support for a malpedia.json config file. This file is searched for
in the following common locations.

- the current working directory + "malpedia.json"
- the current working directory + ".malpedia.json"
- "$HOME/.malpedia.json"
- "/etc/malpedia.json"
- "%APPDATA%\malpedia\malpedia.json" (for windows)

The first file that was found is used. Should a file be found but a
parse error occurs, a corresponding error message is emitted and the
software continues as if no config is present. Consequently no further
locations are searched if a file was found but does not contain valid
json.

If no such file was found, the software attempts to load a config.py as
previously. This is kept for backwards compatibility.

Example config file:

```json
{
    "username": null,
    "password": null,
    "apitoken": "<justanexample>"
}
```

Apitoken authentication is prefered over username password
authentication.